refactor: move WorkoutActiveScreen to dedicated module
