+++ ui/screens/__init__.py
@@ -1,0 +1,5 @@
+"""Screen module exports."""
+
+from .workout_active_screen import WorkoutActiveScreen
+
+__all__ = ["WorkoutActiveScreen"]
